clean up cypress commands
